JNG-3991 Add unit tests and clean the unnecessary lines and entities.
